correct linting errors; move datastore tests to st2common/test/unit/test_datastore.py
