+++ st2actions/st2actions/runners/pythonrunner.py
@@ -74,7 +74,9 @@ class Action(object):
         """
         self.config = config or {}
         self.logger = self._set_up_logger()
-        self.datastore = None # Instance of DatastoreService is added in PythonActionWrapper._get_action_instance
+
+        # Instance of DatastoreService is added in PythonActionWrapper._get_action_instance
+        self.datastore = None
 
     @abc.abstractmethod
     def run(self, **kwargs):
